메인페이지 좋아요 제외 동작 생성

+++ app/src/main/java/com/wewrite/android/api/model/GroupResponse.kt
@@ -13,9 +13,9 @@ data class GroupResponse (
     @Parcelize
     data class GroupData(
         @SerializedName("groupCode") val groupCode: String,
-        @SerializedName("groupId") val groupId: Int,
+        @SerializedName("groupId") val groupId: Long,
         @SerializedName("groupImageUrl") val groupImageUrl: String,
-        @SerializedName("groupMemberCount") val groupMemberCount: Int,
+        @SerializedName("groupMemberCount") val groupMemberCount: Long,
         @SerializedName("groupName") val groupName: String,
     )
 }
